requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
firebase-admin==5.1.0
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.2.0
outcome==1.1.0
packaging==21.2
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wsproto==1.0.0
yarl==1.7.2
